-v flag outputs version and exits 0
[#134041051](https://www.pivotaltracker.com/story/show/134041051)

+++ main/agent.go
@@ -51,6 +51,11 @@ func startAgent(logger logger.Logger) error {
 		return err
 	}
 
+	if opts.VersionCheck {
+		fmt.Println(VersionLabel)
+		os.Exit(0)
+	}
+
 	sigCh := make(chan os.Signal, 8)
 	signal.Notify(sigCh, syscall.SIGTERM, os.Interrupt, os.Kill)
 	errCh := runAgent(opts, logger)
